Finished wrapper for User and according unit tests.
This commit adds support for all public functions of the User class, which allows for accessing the operands of instructions.

What remains is to implement the Use class, that keeps track of all uses of a particular value. The boilerplate for this class is also already included in this commit.